fix(mailgun): move OpenAPI spec to providers/mailgun/openapi
