+++ providers/mailgun/openapi/schema.yaml
